Implemented support for content fragment creation

+++ de.enithing.contenthub.generator.core/src/main/java/de/enithing/contenthub/generator/context/ContextGenerator.java
@@ -1,24 +1,28 @@
-package de.enithing.contenthub.generator;
+package de.enithing.contenthub.generator.context;
 
-import java.awt.datatransfer.StringSelection;
 import java.io.IOException;
 import java.io.OutputStream;
-import java.io.PrintWriter;
+import java.io.StringWriter;
 import java.io.Writer;
 import java.nio.file.Path;
-import java.util.ArrayList;
+import java.util.List;
+import java.util.stream.Collectors;
 
-import org.apache.commons.lang3.StringUtils;
+import org.apache.commons.lang.NotImplementedException;
 import org.apache.commons.vfs2.FileObject;
 import org.apache.velocity.Template;
 import org.apache.velocity.VelocityContext;
 import org.apache.velocity.runtime.parser.ParseException;
 
+import de.enithing.contenthub.generator.GeneratorConfiguration;
+import de.enithing.contenthub.generator.TemplateBasedGenerator;
+import de.enithing.contenthub.generator.contentfragment.instance.ContentFragmentGenerator;
+import de.enithing.contenthub.generator.util.VelocityUtils;
 import de.enithing.contenthub.generator.util.XmlUtils;
+import de.enithing.contenthub.model.contentfragment.AllowedContentFragmentModelPolicy;
 import de.enithing.contenthub.model.contentfragment.ContentFragmentInstance;
-import de.enithing.contenthub.model.contentfragment.ContentFragmentModel;
 import de.enithing.contenthub.model.contenthub.Context;
-import de.enithing.contenthub.model.contenthub.ContextType;
+import de.enithing.contenthub.model.contenthub.ContextPolicy;
 
 public class ContextGenerator implements TemplateBasedGenerator<Context> {
 	private GeneratorConfiguration config;
@@ -37,10 +41,28 @@ public class ContextGenerator implements TemplateBasedGenerator<Context> {
 		return Path.of("/context");
 	}
 
-	@Override
-	public void onEnter(Context ctx) throws Exception {
+	private String renderPolicy(Context ctx, ContextPolicy policy) throws Exception {
+		final VelocityContext templateCtx = getTemplateContext(ctx);
+		Template tpl = resolveTemplate(ctx, "policy.xml");
+
+		StringWriter writer = new StringWriter();
+
+		if (policy instanceof AllowedContentFragmentModelPolicy) {
+			AllowedContentFragmentModelPolicy p = ((AllowedContentFragmentModelPolicy) policy);
+			List<Path> paths = p.getModels().stream().map(mdl -> VelocityUtils.replace(mdl.getPath(), templateCtx))
+					.collect(Collectors.toList());
+			templateCtx.put("policyPaths", paths);
+		} else {
+			throw new NotImplementedException();
+		}
+
+		tpl.merge(templateCtx, writer);
+
+		return writer.toString();
+	}
+	
+	private FileObject getOrCreateTargetRoot(Context ctx, boolean deleteIfExists) throws Exception {
 		FileObject targetRoot;
-		Template tpl = resolveTemplate(ctx, ".content.xml");
 
 		if (ctx.getParentContext() != null) {
 			VelocityContext templateContext = getTemplateContext(ctx);
@@ -49,7 +71,7 @@ public class ContextGenerator implements TemplateBasedGenerator<Context> {
 			String subFolder = VelocityUtils.replace(ctx.getName(), templateContext);
 			targetRoot = getConfig().targetRoot.resolveFile(subFolder);
 
-			if (targetRoot.exists()) {
+			if (deleteIfExists && targetRoot.exists()) {
 				// Clean up
 				targetRoot.deleteAll();
 			}
@@ -58,8 +80,24 @@ public class ContextGenerator implements TemplateBasedGenerator<Context> {
 		} else {
 			targetRoot = getConfig().targetRoot;
 		}
+		
+		return targetRoot;
+	}
+
+	@Override
+	public void onEnter(Context ctx) throws Exception {
+		Template tpl = resolveTemplate(ctx, ".content.xml");
+
+		FileObject targetRoot = getOrCreateTargetRoot(ctx, true);
 
 		VelocityContext templateCtx = getTemplateContext(ctx);
+
+		// Render the policies
+		String renderedPolicies = "";
+		for (ContextPolicy policy : ctx.getPolicies()) {
+			renderedPolicies += renderPolicy(ctx, policy);
+		}
+		templateCtx.put("renderedPolicies", renderedPolicies);
 
 		templateCtx.put("contextTitle", ctx.getTitle());
 		templateCtx.put("contextPrimaryType", ctx.getPrimaryType());
@@ -79,9 +117,8 @@ public class ContextGenerator implements TemplateBasedGenerator<Context> {
 		VelocityContext templateContext = getTemplateContext(ctx);
 
 		// Replace the variables used in the path
-		String subFolder = VelocityUtils.replace(ctx.getName(), templateContext);
-		final FileObject targetRoot = getConfig().targetRoot.resolveFile(subFolder);
-
+		FileObject targetRoot = getOrCreateTargetRoot(ctx, false);
+		
 		// Write the content fragments
 		for (ContentFragmentInstance cf : ctx.getContentFragments()) {
 			GeneratorConfiguration childConfig = createChildConfig(ctx);
